chore: folder refactor
