+++ src/app/create-certificate/page.tsx
@@ -1,4 +1,4 @@
-import CreateCertificatePage from "@/components/templates/create-certificate-page"
+import CreateCertificatePage from "@/components/templates/CreateCertificate/CreateCertificatePage"
 
 export default function CreateCertificate() {
   return <CreateCertificatePage/>
